Cs v1.10.0
-added more AngleSets to TetraTriFlake
-fixed crash when lowering max threads
-fixed Antispin not working correctly in OfDepth mode
-fixed crash when changing the number of children in the editor
-and some more bugfixes
-expanded and custom color palettes
-no zoom option (make sure to account for in hue multiplier)
-fixed OfDepth pixels

+++ Cs/Components/AnimatedGifEncoder.cs
@@ -815,16 +815,21 @@ namespace Gif.Components {
 		public bool Abort(bool fail = true) {
 			if (!fail)
 				return true;
-			started = false;
-			ms?.Close();
-			fs?.Close();
-			if (fs != null && filename != null && filename != "")
-				File.Delete(filename);
-			fs = null;
-			ms = null;
-			finishedAnimation = false;
-			addedFrames = finishedFrame = 0;
-			encodeTaskData = writeTaskData = null;
+			Monitor.Enter(this);
+			try {
+				if (!started)
+					return false;
+				started = false;
+				ms?.Close();
+				fs?.Close();
+				if (fs != null && filename != null && filename != "")
+					File.Delete(filename);
+				fs = null;
+				ms = null;
+				finishedAnimation = false;
+				addedFrames = finishedFrame = 0;
+				encodeTaskData = writeTaskData = null;
+			} finally { Monitor.Exit(this); }
 			return false;
 		}
 		protected TryWrite TryWriteInternal(bool parallel) {
